extracted simple functions, global time variables, minor remove left over logging lines
